moved the examples into kubernetes-examples
and added the extra API check to help the user know if the cluster supports the required API Group or not
