+++ kubernetes-examples/src/main/java/io/fabric8/openshift/examples/DeploymentConfigScale.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.fabric8.openshift.client.examples;
+package io.fabric8.openshift.examples;
 
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.openshift.api.model.DeploymentConfig;
@@ -22,6 +22,7 @@ import io.fabric8.openshift.api.model.DeploymentConfigList;
 import io.fabric8.openshift.api.model.DeploymentConfigSpec;
 import io.fabric8.openshift.api.model.DoneableDeploymentConfig;
 import io.fabric8.openshift.client.DefaultOpenShiftClient;
+import io.fabric8.openshift.client.OpenShiftAPIGroups;
 import io.fabric8.openshift.client.OpenShiftClient;
 import io.fabric8.openshift.client.dsl.DeployableScalableResource;
 
@@ -48,6 +49,10 @@ public class DeploymentConfigScale {
       Integer oldReplicas = 0;
 
       OpenShiftClient client = new DefaultOpenShiftClient();
+      if (!client.supportsOpenShiftAPIGroup(OpenShiftAPIGroups.APPS)) {
+        System.out.println("WARNING this cluster does not support the API Group " + OpenShiftAPIGroups.APPS);
+        return;
+      }
       DeployableScalableResource<DeploymentConfig, DoneableDeploymentConfig> resource = client.deploymentConfigs().withName(name);
       DeploymentConfig deploymentConfig = resource.get();
       if (deploymentConfig == null) {
