+++ kubernetes-examples/src/main/java/io/fabric8/openshift/examples/ListBuildConfigs.java
@@ -14,39 +14,33 @@
  * limitations under the License.
  */
 
-package io.fabric8.openshift.client.examples;
+package io.fabric8.openshift.examples;
 
 import io.fabric8.kubernetes.client.KubernetesClientException;
-import io.fabric8.openshift.api.model.DeploymentConfig;
-import io.fabric8.openshift.api.model.DeploymentConfigList;
+import io.fabric8.openshift.api.model.BuildConfig;
+import io.fabric8.openshift.api.model.BuildConfigList;
 import io.fabric8.openshift.client.DefaultOpenShiftClient;
+import io.fabric8.openshift.client.OpenShiftAPIGroups;
 import io.fabric8.openshift.client.OpenShiftClient;
 
 import java.util.List;
 
-import static org.junit.Assert.assertNotNull;
-
-public class ListDeploymentConfigs {
+public class ListBuildConfigs {
   public static void main(String[] args) {
     try {
       OpenShiftClient client = new DefaultOpenShiftClient();
-      DeploymentConfigList list = client.deploymentConfigs().list();
+      if (!client.supportsOpenShiftAPIGroup(OpenShiftAPIGroups.BUILD)) {
+        System.out.println("WARNING this cluster does not support the API Group " + OpenShiftAPIGroups.BUILD);
+        return;
+      }
+      BuildConfigList list = client.buildConfigs().list();
       if (list == null) {
         System.out.println("ERROR no list returned!");
         return;
       }
-      List<DeploymentConfig> items = list.getItems();
-      for (DeploymentConfig item : items) {
-        System.out.println("DeploymentConfig " + item.getMetadata().getName() + " has version: " + item.getApiVersion());
-      }
-
-      if (items.size() > 0) {
-        // lets check .get() too
-        DeploymentConfig deploymentConfig = items.get(0);
-        String name = deploymentConfig.getMetadata().getName();
-        deploymentConfig = client.deploymentConfigs().withName(name).get();
-        assertNotNull("No DeploymentConfig found for name " + name, deploymentConfig);
-        System.out.println("get() DeploymentConfig " + name + " has version: " + deploymentConfig.getApiVersion());
+      List<BuildConfig> items = list.getItems();
+      for (BuildConfig item : items) {
+        System.out.println("BuildConfig " + item.getMetadata().getName() + " has version: " + item.getApiVersion());
       }
     } catch (KubernetesClientException e) {
       System.out.println("Failed: " + e);
