+++ kubernetes-examples/src/main/java/io/fabric8/openshift/examples/ListImageStreams.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package io.fabric8.openshift.client.examples;
+package io.fabric8.openshift.examples;
 
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.openshift.api.model.ImageStream;
 import io.fabric8.openshift.api.model.ImageStreamList;
 import io.fabric8.openshift.client.DefaultOpenShiftClient;
+import io.fabric8.openshift.client.OpenShiftAPIGroups;
 import io.fabric8.openshift.client.OpenShiftClient;
 
 import java.util.List;
@@ -28,6 +29,10 @@ public class ListImageStreams {
   public static void main(String[] args) {
     try {
       OpenShiftClient client = new DefaultOpenShiftClient();
+      if (!client.supportsOpenShiftAPIGroup(OpenShiftAPIGroups.IMAGE)) {
+        System.out.println("WARNING this cluster does not support the API Group " + OpenShiftAPIGroups.IMAGE);
+        return;
+      }
       ImageStreamList list = client.imageStreams().list();
       if (list == null) {
         System.out.println("ERROR no list returned!");
